Trabalho_CC2/requirements.txt
Django==3.2.3
django-allauth==0.44.0
django-environ==0.4.5
django-filter==2.4.0
django-rest-framework==0.1.0
djangorestframework==3.12.4
markdown==3.3.4
mysqlclient==2.0.3
